+++ plugins/module_utils/__init__.py
@@ -36,9 +36,13 @@ from ansible_collections.o0_o.utils.plugins.module_utils.date_utils import (  # 
     format_epoch_timestamp,
     parse_date_to_epoch,
     parse_datetime,
+    parse_elapsed_time,
 )
 from ansible_collections.o0_o.utils.plugins.module_utils.string_utils import (  # noqa: E501
     strip_comments,
+)
+from ansible_collections.o0_o.utils.plugins.module_utils.lookup_utils import (  # noqa: E501
+    VarsLookupBase,
 )
 
 __all__ = [
@@ -54,5 +58,7 @@ __all__ = [
     "format_epoch_timestamp",
     "parse_date_to_epoch",
     "parse_datetime",
+    "parse_elapsed_time",
     "strip_comments",
+    "VarsLookupBase",
 ]